+++ TMS-backend/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "tms-backend",
+  "version": "1.0.0",
+  "description": "backend",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "group4",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.17.3"
+  }
+}
