+++ TMS-frontend/src/component/pages/footer.css
@@ -1,0 +1,10 @@
+.main-footer{
+    color: white;
+    background-color:gray;
+    padding-top: 2cm;
+    position: relative;
+    bottom: 0;
+    width: 100%;
+    margin-top: 450px;
+}
+
