--- src/component/pages/about.css
@@ -1,16 +1,0 @@
-.showcase1 {
-    background: url('./images/about.jpg') no-repeat center/cover;
-    height: 500px;
-    width: 100%;
-    color: white;
-    padding-top: 90px;
-}
-
-.box-4{
-    font-size: 36px;
-    padding-bottom: 20px; 
-    padding-top: 60px;
-    margin-top: 50px;
-    height: 30px;
- 
-}
