--- src/component/pages/errorpage.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-
-const ErrorPage = ()=> {
-    return (
-        <div className="container">
-            <div className="py-4">
-                <h1>ErrorPage not found</h1>
-            </div>
-            </div>
-      
-    );
-};
-export default ErrorPage
