--- src/component/pages/home.css
@@ -1,49 +1,0 @@
-.showcase {
-    background: url('./images/home.jpg') no-repeat center/cover;
-    height: 600px;
-    width: 100%;
-    color: black;
-
-}
-/* .showcase.overlay { */
-    .box-1{
-    display: flex;
-    align-items: center;
-    height: 600px;
-    justify-content: center;
-    flex-direction: column;
-    color:white;
-    background-color: rgba(0,0,0,0.4);
-    
-}
-/* .showcase.overlay { */
-    .box-1{
-   font-size: 36px;
-   padding-bottom: 20px; 
-
-}
-/* .box {
-    background-color:;
-   height: 150px;
-} */
-.box-2{
-    display: flex;
-    align-items: center;
-    height: 80px;
-    justify-content: center;
-    /* flex-direction: column; */
-   
-    /* color:white; */
-    /* background-color: rgba(0,0,0,0.4); */
-
-}
-.box-3{
-    align-items: center;
-    height: 250px;
-    margin-right: 35px;
-}
-.Button{
-    margin-top: 5px;
-    height: 200px;
- 
-}
