--- src/component/pages/login.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-
-const Login = ()=> {
-    return (
-        <div className="container">
-            <div className="py-4">
-                <h1>Login Page</h1>
-            </div>
-       </div>
-    );
-};
-export default Login
